+++ src/main/java/com/example/bookhub/admin/controller/CategoryController.java
@@ -1,0 +1,39 @@
+package com.example.bookhub.admin.controller;
+
+import com.example.bookhub.admin.service.CategoryService;
+import com.example.bookhub.admin.vo.Category;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+@RequestMapping("/admin")
+@RequiredArgsConstructor
+public class CategoryController {
+
+    private final CategoryService categoryService;
+
+    @GetMapping("/category")
+    public String category(Model model) {
+        List<Category> topLevelCategories = categoryService.getAllTopLevelCategories();
+        List<Category> secondLevelCategories = categoryService.getAllSecondLevelCategories();
+        List<Category> thirdLevelCategories = categoryService.getAllThirdLevelCategories();
+
+        model.addAttribute("topLevelCategories", topLevelCategories);
+        model.addAttribute("secondLevelCategories", secondLevelCategories);
+        model.addAttribute("thirdLevelCategories", thirdLevelCategories);
+
+        return "admin/category";
+    }
+
+    @GetMapping("/secondCategory")
+    @ResponseBody
+    public List<Category> handleRequest(@RequestParam("category") int categoryNo) {
+        return categoryService.getSecondLevelCategoriesByTopLevelCategoryName(categoryNo);
+    }
+}
