+++ src/main/resources/mybatis/mappers/Admin/Category.xml
@@ -29,6 +29,16 @@
         							FROM BookHub.CATEGORY
         							WHERE PARENT_CATEGORY_NO IN (SELECT CATEGORY_NO
         														FROM BookHub.CATEGORY
-        														WHERE PARENT_CATEGORY_NO IS NULL));
+        														WHERE PARENT_CATEGORY_NO IS NULL))
+    </select>
+
+    <select id="getSecondLevelCategoriesByTopLevelCategoryName" parameterType="int" resultType="com.example.bookhub.admin.vo.Category">
+            SELECT CATEGORY_NO AS no,
+            		CATEGORY_NAME AS name,
+                    PARENT_CATEGORY_NO AS parentNo
+            FROM BookHub.CATEGORY
+            WHERE PARENT_CATEGORY_NO IN (SELECT CATEGORY_NO
+            							FROM BookHub.CATEGORY
+            							WHERE CATEGORY_NO = #{value})
     </select>
 </mapper>
